delete app esg
